+++ src/main/java/com/github/mkram17/bazaarutils/features/Bookmark.java
@@ -100,7 +100,7 @@ public class Bookmark extends CustomItemButton {
             return null;
         OrderPriceInfo priceInfo = new OrderPriceInfo(0.0, OrderPriceInfo.priceTypes.INSTABUY);
 
-        priceInfo.updateMarketPrice(BazaarData.findProductId(name));
+        priceInfo.updateMarketPrices(BazaarData.findProductId(name));
         if (priceInfo.getPricePerItem() == 0.0 && priceInfo.getMarketPrice() == 0.0) {
             PlayerActionUtil.notifyAll("Could not find prices for " + name + ", try to bookmark it again.", Util.notificationTypes.BAZAARDATA);
             return null;
