+++ src/main/java/com/github/mkram17/bazaarutils/utils/BUCommands.java
@@ -3,8 +3,8 @@ package com.github.mkram17.bazaarutils.utils;
 import com.github.mkram17.bazaarutils.config.BUConfig;
 import com.github.mkram17.bazaarutils.features.CustomOrder;
 import com.github.mkram17.bazaarutils.features.OutdatedOrderHandler;
-import com.github.mkram17.bazaarutils.features.restrictsell.RestrictSell;
-import com.github.mkram17.bazaarutils.features.restrictsell.RestrictSellControl;
+import com.github.mkram17.bazaarutils.features.restrictsell.InstaSellRestrictions;
+import com.github.mkram17.bazaarutils.features.restrictsell.SellRestrictionControl;
 import com.github.mkram17.bazaarutils.misc.orderinfo.OrderData;
 import com.mojang.brigadier.CommandDispatcher;
 import com.mojang.brigadier.arguments.DoubleArgumentType;
@@ -172,7 +172,7 @@ public class BUCommands {
                                 .then(ClientCommandManager.argument("limit", DoubleArgumentType.doubleArg(0.1))
                                         .executes(context -> {
                                             double limit = DoubleArgumentType.getDouble(context, "limit");
-                                            BUConfig.get().restrictSell.addRule(RestrictSell.restrictBy.VOLUME, limit);
+                                            BUConfig.get().instaSellRestrictions.addRule(InstaSellRestrictions.restrictBy.VOLUME, limit);
                                             Util.scheduleConfigSave();
                                             PlayerActionUtil.notifyAll("Added rule: VOLUME" + ": " + limit);
                                             return 1;
@@ -184,7 +184,7 @@ public class BUCommands {
                                 .then(ClientCommandManager.argument("limit", DoubleArgumentType.doubleArg(0.1))
                                         .executes(context -> {
                                             double limit = DoubleArgumentType.getDouble(context, "limit");
-                                            BUConfig.get().restrictSell.addRule(RestrictSell.restrictBy.PRICE, limit);
+                                            BUConfig.get().instaSellRestrictions.addRule(InstaSellRestrictions.restrictBy.PRICE, limit);
                                             Util.scheduleConfigSave();
                                             PlayerActionUtil.notifyAll("Added rule: PRICE" + ": " + limit);
                                             return 1;
@@ -196,7 +196,7 @@ public class BUCommands {
                                 .then(ClientCommandManager.argument("itemName", StringArgumentType.string())
                                         .executes(context -> {
                                             String name = StringArgumentType.getString(context, "itemName");
-                                            BUConfig.get().restrictSell.addRule(RestrictSell.restrictBy.NAME, name);
+                                            BUConfig.get().instaSellRestrictions.addRule(InstaSellRestrictions.restrictBy.NAME, name);
                                             Util.scheduleConfigSave();
                                             PlayerActionUtil.notifyAll("Added rule: NAME" + ": " + name);
                                             return 1;
@@ -210,13 +210,13 @@ public class BUCommands {
                         .then(ClientCommandManager.argument("rule number", IntegerArgumentType.integer(1))
                                 .executes(context -> {
                                     int restrictNum = IntegerArgumentType.getInteger(context, "rule number") - 1;
-                                    RestrictSellControl rule = BUConfig.get().restrictSell.getControls().get(restrictNum);
+                                    SellRestrictionControl rule = BUConfig.get().instaSellRestrictions.getControls().get(restrictNum);
                                     if (rule == null)
                                         context.getSource().sendError(Text.literal("Invalid rule number. Check the order in /bu"));
                                     if (rule.getRule() != null) {
-                                        PlayerActionUtil.notifyAll(rule.getRule() == RestrictSell.restrictBy.NAME ? "Removed rule: NAME: " + rule.getName() : "Removed rule: " + rule.getRule() + ": " + rule.getAmount());
+                                        PlayerActionUtil.notifyAll(rule.getRule() == InstaSellRestrictions.restrictBy.NAME ? "Removed rule: NAME: " + rule.getName() : "Removed rule: " + rule.getRule() + ": " + rule.getAmount());
                                     }
-                                    BUConfig.get().restrictSell.getControls().remove(restrictNum);
+                                    BUConfig.get().instaSellRestrictions.getControls().remove(restrictNum);
                                     Util.scheduleConfigSave();
                                     return 1;
                                 })
