+++ src/main/java/com/github/mkram17/bazaarutils/utils/instasell/InstaSellItem.java
@@ -1,0 +1,5 @@
+package com.github.mkram17.bazaarutils.utils.instasell;
+
+import com.github.mkram17.bazaarutils.misc.orderinfo.OrderPriceInfo;
+
+public record InstaSellItem(String name, OrderPriceInfo priceInfo){}
